feat(upgrade): split dependency upgrade from configuration update  (#343)
Currently the `upgrade-dependencies` workflow includes both dependencies, as as project configuration (controlled by projen). This is problematic because sometimes it causes a conflict: A project configuration needs to change in order to upgrade its dependencies (for example when dependencies no longer support a specific node version).

This PR splits the workflow into two:

- One for dependencies upgrade, named `upgrade-dependencies`
- One for configuration upgrade, named `upgrade-configuration`.