chore: offset timing of runtime dependency upgrades
We are seeing failures in these actions because they are on the same schedule as
dev dependency upgrades. The dev dependency upgrades finish slightly faster and cause these to fail
because the version required for cdk8s is higher than the depdency upgrade in this workflow.

See https://github.com/cdk8s-team/cdk8s-postgres/actions/runs/5660984217/job/15337905770.

Signed-off-by: Kendra Neil <53584728+TheRealAmazonKendra@users.noreply.github.com>