fix: `downlevel-dts` package causes daily identical releases (#528)
There's no real need to bump the typescript version used by downlevel-dts. Its also unclear why they depend on a nightly unstable typescript, and we are better off operating on the stable typescript version. 

Not that all of this matters anyway because we don't emit downleveled types anyway (because we dont need to), and its unlikely we ever will.