fix: isolated peer upgrades break upgrade workflows
As for example seen in: https://github.com/cdk8s-team/cdk8s-hasura/pull/407

The workflow upgrades the peer dependency:

```
cdk8s-plus-27: ^2.7.61 -> ^2.7.66
```

But leaves the devDependency unchanged:

```
cdk8s-plus-27: 2.7.61
```

And then the install fails because its requirements are not satisfied.

Instead, don't upgrade peer dependencies automtically. If we need to do
a peerDependency upgrade, we will do it by hand.

Signed-off-by: Rico Huijbers <rix0rrr@gmail.com>