fix(node): projen upgrades not working
This might actually a bug in ncu, but with the previous configuration ncu only proposed to update the peer dependency constraint to`^0` (which is later reverted by projen) and does not attempt to update the dev dep constraint.

This fix makes `ncu` not look at the projen peer dependency, because it is hardcoded anyway.

Signed-off-by: Momo Kornher <mail@moritzkornher.de>