fix(node): projen upgrades not working (#760)
This might actually a bug in ncu, but with the previous configuration ncu only proposed to update the peer dependency constraint to`^0` (which is later reverted by projen) and does not attempt to update the dev dep constraint.

This fix makes `ncu` not look at the projen peer dependency, because it is hardcoded anyway.

Tested locally on a previous commit:

<img width="858" alt="image" src="https://github.com/cdk8s-team/cdk8s-projen-common/assets/379814/5cac6c8b-8862-445c-a1f2-c620b571ac9c">