chore: centralize issues in `cdk8s` repository (#961)
A while back we had a [discussion](https://github.com/cdk8s-team/cdk8s/discussions/1274) on whether to have issues only in the `cdk8s` repository with appropriate labels, as opposed to scattered across 3 repos. Quick poll on slack showed the community is in favor, and I have no objections - so lets do it.