feat: implement backport using native `projen` components (#1051)
We are seeing a several security alerts that originate from the
`backport` dependency.
 
- https://github.com/cdk8s-team/cdk8s-cli/issues/2825
- https://github.com/cdk8s-team/cdk8s-core/issues/2932
- https://github.com/cdk8s-team/cdk8s-plus/issues/4761 

We could fix it here, but lets migrate to the native projen
implementation and if needed, fix in projen.

### Testing

Its tricky to test if this breaks backport for our use case is some
subtle way. Thats fine, if something breaks we'll just fix it.

Signed-off-by: epolon <epolon@amazon.com>